Move AttrMap functionality to 'internal/create'.
